requirements.txt
tf-keras
tensorflow
transformers
fastapi
uvicorn
pillow
numpy
pydantic
python-multipart
sqlalchemy
psycopg2-binary
python-decouple
opencv-python
mediapipe
supabase
python-dotenv
scikit-learn
python-multipart
pydantic-settings
loguru